Relocate conversion functions for unofficial endplints
